Add files via upload
First Review Presentation containing a description of what the project is all about. It also explains about the data structure used and it's upper hand over other data structures. Definition of classes that will be used in the main project along with the implementation of Binary Search Tree has also been added.